requirements.txt
diffusers==0.30.2
accelerate==0.28.0
numpy==1.23.5
tensorflow==2.12.0
tensorboard==2.12.0
opencv-python==4.9.0.80
soundfile==0.12.1
transformers==4.39.2
huggingface_hub==0.30.2
librosa==0.11.0
einops==0.8.1
gradio==5.24.0
gdown
requests
imageio[ffmpeg]
omegaconf
ffmpeg-python
moviepy
fastapi==0.115.5
websockets==14.1
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
av==13.1.0
aiofiles==24.1.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
typing-extensions==4.12.2